+++ app/assets/stylesheets/opportunities.css
@@ -1,0 +1,148 @@
+* { box-sizing: border-box; }
+
+body {
+  background-color: #fff;
+  color: #333;
+  margin: 0;
+}
+
+body, p, ol, ul, td {
+  font-family: verdana, arial, helvetica, sans-serif;
+  font-size: 13px;
+  line-height: 18px;
+}
+
+pre {
+  background-color: #eee;
+  font-size: 11px;
+  overflow-x: auto;
+  padding: 10px;
+  white-space: pre-wrap;
+  white-space: -moz-pre-wrap;
+  white-space: -pre-wrap;
+  white-space: -o-pre-wrap;
+  word-wrap: break-word;
+}
+
+article { margin: 10px 33px; }
+
+header,
+footer {
+  background-color: grey;
+  padding: 5px;
+}
+
+header h1 {
+  margin: 0;
+  padding: 0;
+  font-size: 0.8em;
+  color: white;
+}
+
+header a,
+footer a {
+  text-decoration: none;
+  color: white;
+}
+header a:visited,
+footer a:visited {
+  color: white;
+}
+header a:hover,
+footer a:hover {
+  background-color: transparent;
+  text-decoration: underline;
+}
+
+footer { font-size: .8em; }
+
+a { color: #000; }
+a:visited { color: #666; }
+
+a:hover {
+  color: #fff;
+  background-color: #000;
+}
+
+td {
+  border: 1px solid grey;
+  padding: 0 5px 7px;
+}
+tr td:first-child {
+  background-color: #eee;
+  font-weight: 900;
+}
+
+ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+form { display: flex; }
+
+label { display: block; }
+
+h1 { line-height: 1.1em; }
+
+input[type=text] {
+  border: 1px solid grey;
+  padding: 8px;
+  margin: 0px 5px;
+  font-size: 1.1em;
+  border-radius: 3px;
+}
+
+input[type=submit] {
+  background-color: green;
+  background-image: linear-gradient(lightgreen, green);
+  border: 1px solid grey;
+  border-radius: 5px;
+  font-size: 1.2em;
+  padding: 0px 5px;
+  margin: 0;
+}
+input[type=submit]:hover {
+  color: white;
+}
+
+.warning {
+  margin: 1em;
+  color: white;
+  background: grey;
+  font-size: 2em;
+  border: 1px solid black;
+  border-radius: 5px;
+  padding: 1em;
+  line-height: 1.1em;
+}
+
+.procurement-short {
+  border: 1px solid grey;
+  border-radius: 3px;
+  box-shadow: 0 2px 2px 0 rgba(0,0,0,0.14),
+    0 3px 1px -2px rgba(0,0,0,0.12),
+    0 1px 5px 0 rgba(0,0,0,0.2);
+  margin: 1em 0;
+  padding: 5px;
+}
+
+.procurement-short h2 {
+  font-size: 1.1em;
+  margin: 0;
+  margin-top: .5em;
+}
+
+.procurement-short h1 {
+  margin: 0;
+  margin-top: .5em;
+}
+
+.procurement-short .buyer-org {
+  margin: 0;
+  margin-right: .5em;
+  margin-bottom: 1em;
+  padding: 0;
+  font-size: 1.1em;
+  color: grey;
+  border-bottom: 1px solid #ddd;
+}
