+++ app/controllers/api/api_controller.rb
@@ -1,0 +1,13 @@
+module Api
+
+class ApiController < ActionController::API
+ before_action :set_default_response_format
+
+ private
+
+ def set_default_response_format
+   request.format = :json
+ end
+end
+
+end  # module Api
