+++ app/controllers/application_controller.rb
@@ -1,9 +1,2 @@
-class ApplicationController < ActionController::API
-  before_action :set_default_response_format
-
-  private
-
-  def set_default_response_format
-    request.format = :json
-  end
+class ApplicationController < ActionController::Base
 end
